Fix creation of tables in database connection

+++ code/fwc/src/main/java/com/elementaryengineers/fwc/panel/StudentHistory.java
@@ -112,16 +112,22 @@ public class StudentHistory extends JPanel {
 
             // Check if selected a worksheet
             if (index > 0) {
-                // Check database update status
-                if (FWCConfigurator.getDbConn().deleteWorksheet(sheets.get(index).getWorksheetID())) {
-                    sheets.remove(index); // Remove from student history
-                    populateTable(); // Refresh table of worksheets
-                    JOptionPane.showMessageDialog(null, "Worksheet has been deleted successfully!",
-                            "Delete Worksheet Successful", JOptionPane.INFORMATION_MESSAGE);
-                }
-                else {
-                    JOptionPane.showMessageDialog(null, "Worksheet could not be deleted! Please try again later.",
-                            "Delete Worksheet Failed", JOptionPane.INFORMATION_MESSAGE);
+                int confirm = JOptionPane.showConfirmDialog(null,
+                        "Are you sure you want to delete this worksheet?",
+                        "Delete Confirmation", JOptionPane.YES_NO_OPTION);
+
+                if (confirm == JOptionPane.YES_OPTION) {
+                    // Check database update status
+                    if (FWCConfigurator.getDbConn().deleteWorksheet(sheets.get(index).getWorksheetID())) {
+                        sheets.remove(index); // Remove from student history
+                        populateTable(); // Refresh table of worksheets
+                        JOptionPane.showMessageDialog(null, "Worksheet has been deleted successfully!",
+                                "Delete Worksheet Successful", JOptionPane.INFORMATION_MESSAGE);
+                    }
+                    else {
+                        JOptionPane.showMessageDialog(null, "Worksheet could not be deleted! Please try again later.",
+                                "Delete Worksheet Failed", JOptionPane.INFORMATION_MESSAGE);
+                    }
                 }
             }
             else { // No worksheet is selected from the table
